factored commandline parser for use in compiled context
